Updated to Arduino 1.5.x+ library specification.

+++ examples/SavedVariableArray/SavedVariableArray.ino
@@ -1,11 +1,12 @@
-#include "Arduino.h"
-#include "Streaming.h"
+#include <Arduino.h>
+#include <Streaming.h>
 #ifndef ARDUINO_SAM_DUE
 #include <EEPROM.h>
 #endif
-#include "SavedVariable.h"
+#include <SavedVariable.h>
 
-const size_t BAUDRATE = 115200;
+
+const long BAUDRATE = 115200;
 
 const size_t EEPROM_INITIALIZED_VALUE = 44;
