+++ examples/SavedVariableConstantString/SavedVariableConstantString.ino
@@ -1,13 +1,13 @@
-#include "Arduino.h"
-#include "Streaming.h"
+#include <Arduino.h>
+#include <Streaming.h>
 #ifndef ARDUINO_SAM_DUE
 #include <EEPROM.h>
 #endif
-#include "SavedVariable.h"
-#include "ConstantVariable.h"
+#include <SavedVariable.h>
+#include <ConstantVariable.h>
 
 
-const size_t BAUDRATE = 115200;
+const long BAUDRATE = 115200;
 
 union SubsetMemberType
 {
